% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLD_trends.R
\name{CLD_trends}
\alias{CLD_trends}
\title{Catch, discards, and landings by stock}
\usage{
CLD_trends(x)
}
\arguments{
\item{x}{a dataframe output of format_sag function}
}
\value{
A data frame..
}
\description{
Returns a data frame of reference points, catch, discards, and landings by stock over time.
}
\note{
Can add some helpful information here
}
\examples{
\dontrun{
CLD_2018 <- catch_landings_discards(sag_formatted)
}

}
\references{
The ICES stock information Database web sevices: \url{http://sid.ices.dk/services/}
}
\seealso{
\code{\link{stock_trends}} for formatting data from the ICES Stock Assessment database. 

\code{\link{icesFO-package}} gives an overview of the package.
}
