% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_name.r
\name{file_name}
\alias{file_name}
\title{Helps creating the name of the file to save and save it in the right folder}
\usage{
file_name(
  year,
  ecoregion,
  name,
  ext = "",
  dir = c("bootstrap", "data", "model", "report")
)
}
\arguments{
\item{year}{the active year}

\item{ecoregion}{the relevant ices ecoregion}

\item{name}{(name of the file)}

\item{ext}{(extention of the file, for example: csv, png...)}

\item{dir}{(directory where to save the file)}
}
\value{
A data frame
}
\description{
Helps creating the name of the file to save and save it in the right folder
}
\note{
Can add some helpful information here
}
\examples{
\dontrun{
sid_format <- format_sid(x, "Celtic Seas")
}

}
\seealso{
\code{\link{format_sag}} for formatting raw data from the ICES Stock Assessment database.

\code{\link{icesFO-package}} gives an overview of the package.
}
