% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_annex_table.R
\name{format_annex_table}
\alias{format_annex_table}
\title{Format the SAG status data for its use as an annex table.}
\usage{
format_annex_table(df, year)
}
\arguments{
\item{df}{a dataframe output from format_sag_status.}

\item{year}{the active year}
}
\value{
A data frame
}
\description{
It provides an excel file with color names per status, as well as
other information that might be required for an annex table
}
\note{
Can add some helpful information here
}
\examples{
\dontrun{
sid_format <- format_sid(x, "Celtic Seas")
}

}
\seealso{
\code{\link{format_sag}} for formatting raw data from the ICES Stock Assessment database.

\code{\link{icesFO-package}} gives an overview of the package.
}
