% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_annex_table_html.r
\name{format_annex_table_html}
\alias{format_annex_table_html}
\title{It provides an html file with color symbols corresponding to fishstock status, as well as
other information that might be required for an annex table
After saving the table in html, open the file in browser. press ctrl+a and ctrl+c, and 
then paste the table either in a word document or in an xls spreadsheet.}
\usage{
format_annex_table_html(df, ecoregion, year)
}
\arguments{
\item{df}{a dataframe output from format_sag_status.}

\item{year}{the active year}
}
\value{
A data frame
}
\description{
It provides an html file with color symbols corresponding to fishstock status, as well as
other information that might be required for an annex table
After saving the table in html, open the file in browser. press ctrl+a and ctrl+c, and 
then paste the table either in a word document or in an xls spreadsheet.
}
\note{
Can add some helpful information here
}
\examples{
\dontrun{
html_annex_table <- html_annex_table(df, ecoregion, year)
}

}
\seealso{
\code{\link{format_sag}} for formatting raw data from the ICES Stock Assessment database.

\code{\link{icesFO-package}} gives an overview of the package.
}
