% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_catches.R
\name{format_catches}
\alias{format_catches}
\title{The previous function for formatting catches}
\usage{
format_catches(
  year,
  ecoregion,
  historical,
  official,
  preliminary = NULL,
  species_list,
  sid
)
}
\arguments{
\item{year}{}

\item{ecoregion}{}

\item{historical}{}

\item{official}{}

\item{preliminary}{}

\item{species_list}{}

\item{sid}{}
}
\description{
The previous function for formatting catches
}
