% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_stecf.R
\name{format_stecf_effort}
\alias{format_stecf_effort}
\title{Format the data from STECF effort and landings}
\usage{
format_stecf_effort(x)
}
\arguments{
\item{x}{the name of the dataframe with effort data}

\item{y}{the name of the dataframe with landings data}

\item{path, }{the path to the files}
}
\value{
a data frame of stock status relative to reference points and catch, discards, and landings
by stock for the most recent assessment.
}
\description{
Format the data from STECF effort and landings for the specific Ecoregion
for which you are producing the Fisheries Overviews.
}
\details{
These two dataframes have to be downloaded by hand and put in the data folder. 
The proper Annexes have to decided by the user.
}
\note{
Some considerable errors have been identified in the STECF data. Finland and Estonia effort data are not reliable,
and Germany recorded an erroneous haul in 2013. These values have been removed.

Can add some helpful information here
}
\examples{
\dontrun{
stecf_formatted <- format_stecf("Celtic Seas")
}

}
\references{
STECF dissemination tool https://stecf.jrc.ec.europa.eu/web/stecf/dd/effort/graphs-annex
}
\seealso{
\code{\link{format_sag}} for formatting raw data from the ICES Stock Assessment database. 

\code{\link{icesFO-package}} gives an overview of the package.
}
