% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guild_status_trends.R
\name{guild_trends}
\alias{guild_trends}
\title{Wrangling of format_sag output}
\usage{
guild_trends(x)
}
\arguments{
\item{x}{a dataframe output of format_sag function}
}
\value{
A data frame..
}
\description{
Wrangling of format_sag output to obtain a dataframe with time-series of F, Fmsy, SSB and MSY B trigger for
each guild in the Ecoregion, according to the last assessment (relative to the set year)
This guild grouping is used in the Ecosystem Overviews.
}
\note{
Can add some helpful information here
}
\examples{
\dontrun{
trends <- guild_trends(sag_formatted)
}

}
\references{
The ICES stock information Database web sevices: \url{http://sid.ices.dk/services/}
}
\seealso{
\code{\link{format_sag}} for formatting data from the ICES Stock Assessment database. 

\code{\link{icesFO-package}} gives an overview of the package.
}
