% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icesFO-package.R
\docType{package}
\name{icesFO-package}
\alias{icesFO-package}
\alias{icesFO}
\title{Functions to support the creation of ICES Fisheries Overviews}
\description{
Functions to support the creation of ICES Fisheries Overviews, ....
}
\details{
\emph{Loading data:}
\tabular{ll}{
  \code{\link{load_sid}}          \tab load data from SID\cr
}
\emph{Tables:}
\tabular{ll}{
  \code{\link{div}}      \tab divide column values\cr
}
\emph{Plots:}
\tabular{ll}{
  \code{\link{lim}}        \tab compute axis limits\cr
}
}
\author{
Adriana Villamor and Scott Large.
}
