% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_areas.R
\name{load_areas}
\alias{load_areas}
\title{Download ICES areas polygons}
\usage{
load_areas()
}
\value{
A simple features collection
}
\description{
Returns a simple features object with polygons for all
subdivisions
}
\examples{
\dontrun{
  areas <- load_areas()
}

}
\seealso{
\code{\link{icesFO-package}} gives an overview of the package.
}
