% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_asfis_species.R
\name{load_asfis_species}
\alias{load_asfis_species}
\title{Download ASFIS Species list from FAO.}
\usage{
load_asfis_species()
}
\value{
A data frame
}
\description{
Download
}
\examples{
\dontrun{
catches_hist_raw <- load_historical_catches()
catches_official_raw <- load_official_catches()
catches_prelim_raw <- load_preliminary_catches()
}

}
\seealso{
\code{\link{load_sag}} for loading data from the ICES Stock Assessment database. 

\code{\link{icesFO-package}} gives an overview of the package.
}
