% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_STECF.R, R/load_catches.R
\name{load_catches}
\alias{load_catches}
\alias{load_STECF}
\alias{load_historical_catches}
\alias{load_official_catches}
\alias{load_preliminary_catches}
\title{Download catches and effort by country data from STECF dissemantion data web.}
\usage{
load_STECF()

load_historical_catches()

load_official_catches()

load_preliminary_catches(year)
}
\arguments{
\item{year}{the year for which data is required.}
}
\value{
A data frame..

A data frame..
}
\description{
This is a new format from 2015

Download historical and current catches data from the ICES web services
also preliminary catches can be downloaded for the current year
}
\note{
Can add some helpful information here

Can add some helpful information here
}
\examples{
\dontrun{
catches_hist_raw <- load_historical_catches()
catches_official_raw <- load_official_catches()
catches_prelim_raw <- load_preliminary_catches()
}

\dontrun{
catches_hist_raw <- load_historical_catches()
catches_official_raw <- load_official_catches()
catches_prelim_raw <- load_preliminary_catches()
}

}
\references{
STECF data dissemination site \url{https://stecf.jrc.ec.europa.eu/dd/fdi}

The ICES stock information Database web sevices: \url{http://sid.ices.dk/services/}
}
\seealso{
\code{\link{load_sag}} for loading data from the ICES Stock Assessment database. 

\code{\link{icesFO-package}} gives an overview of the package.

\code{\link{load_sag}} for loading data from the ICES Stock Assessment database.

\code{\link{icesFO-package}} gives an overview of the package.
}
