% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ecoregion.R
\name{load_ecoregion}
\alias{load_ecoregion}
\title{Download ecoregion polygons}
\usage{
load_ecoregion(ecoregion, precision = 3)
}
\arguments{
\item{ecoregion}{an ICES ecoregion to download (e.g "Baltic Sea")}

\item{precision}{the numnber of decimal places required in the coordinates}
}
\value{
A simple features collection
}
\description{
Returns a simple features object with a polygon for each
ecoregion
}
\examples{
\dontrun{
  ecoregion <- load_ecoregion("Baltic Sea")
}

}
\seealso{
\code{\link{icesFO-package}} gives an overview of the package.
}
