% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_catches.R
\name{load_historical_catches}
\alias{load_historical_catches}
\title{Download catch data from ICES web services.}
\usage{
load_historical_catches()
}
\arguments{
\item{year}{the year for which data is required.}
}
\value{
A data frame..
}
\description{
Download historical and current catches data from the ICES web services
also preliminary catches can be downloaded for the current year
}
\note{
Can add some helpful information here
}
\examples{
\dontrun{
catches_hist_raw <- load_historical_catches()
catches_official_raw <- load_official_catches()
catches_prelim_raw <- load_preliminary_catches()
}

}
\references{
The ICES stock information Database web sevices: \url{http://sid.ices.dk/services/}
}
\seealso{
\code{\link{load_sag}} for loading data from the ICES Stock Assessment database. 

\code{\link{icesFO-package}} gives an overview of the package.
}
