% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sid.R
\name{load_sid}
\alias{load_sid}
\title{Download data from the ICES Stock Information Database.}
\usage{
load_sid(year)
}
\arguments{
\item{year}{the year for which data is required.}
}
\value{
A data frame..
}
\description{
Download data from the ICES Stock Information Database for the year in
which you are producing the Fisheries Overviews.
}
\note{
Can add some helpful information here
}
\examples{
\dontrun{
sid_raw <- load_sid(2019)
}

}
\references{
The ICES stock information Database web sevices: \url{http://sid.ices.dk/services/}
}
\seealso{
\code{\link{load_sag}} for loading data from the ICES Stock Assessment database. 

\code{\link{icesFO-package}} gives an overview of the package.
}
