% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_statrec2ecoregions.R
\name{load_statrec2ecoregions}
\alias{load_statrec2ecoregions}
\title{Download a statistical rectangles mapped to ecoregion}
\usage{
load_statrec2ecoregions()
}
\value{
A data frame
}
\description{
Returns a data frame of giving the ecoregion for each ICES
statistical rectangle
}
\examples{
\dontrun{
  statrec2ecoregion <- load_statrec2ecoregion()
}

}
\seealso{
\code{\link{icesFO-package}} gives an overview of the package.
}
