% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_catch_trends.R
\name{plot_catch_trends}
\alias{plot_catch_trends}
\title{Returns an ordered plot of catch bars colored according to 
F/F<sub>MSY</sub> and SSB/MSY B<sub>trigger</sub>
by fish category and ecoregion}
\usage{
plot_catch_trends(x, type = c("COMMON_NAME", "COUNTRY", "Guild")[1],
  line_count = 10, plot_type = c("line", "area")[1],
  official_catches_year = 2018, return_data = F)
}
\arguments{
\item{x}{a dataframe output of stockstatus+catch_current.R}

\item{return_data}{a parameter indicating if the data behind the plot should be returned as a dataframe}

\item{guild}{an identifier of the Fisheries guild to plot}

\item{cap_month}{the month to be shown in the figure caption, the accession date to SAG usually}

\item{cap_year}{the year to be shown in the figure caption}
}
\value{
A plot
}
\description{
Wrangling of format_sag output to obtain a dataframe with time-series of F, Fmsy, SSB and MSY B trigger for
each stock in the Ecoregion, according to the last assessment (relative to the set year)
}
\note{
Can add some helpful information here
}
\examples{
\dontrun{
plot1 <- plot_CLD_bar(CLDstatus, guild="Demersal", return_data = TRUE)
}

}
\references{
The ICES stock information Database web sevices: \url{http://sid.ices.dk/services/}
}
\seealso{
\code{\link{plot_CLD_bar}} Stock status relative to reference points. 

\code{\link{icesFO-package}} gives an overview of the package.
}
