% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_catch_current.R
\name{plot_CLD_bar}
\alias{plot_CLD_bar}
\title{Returns an ordered plot of catch bars colored according to 
F/F<sub>MSY</sub> and SSB/MSY B<sub>trigger</sub>
by fish category and ecoregion}
\usage{
plot_CLD_bar(
  x,
  guild,
  caption = TRUE,
  cap_year,
  cap_month,
  return_data = FALSE
)
}
\arguments{
\item{x}{a dataframe output of stockstatus+catch_current.R}

\item{guild}{an identifier of the Fisheries guild to plot}

\item{caption}{logical flag}

\item{cap_year}{the year to be shown in the figure caption}

\item{cap_month}{the month to be shown in the figure caption, the accession date to SAG usually}

\item{return_data}{a parameter indicating if the data behind the plot should be returned as a dataframe}
}
\value{
A plot
}
\description{
Wrangling of format_sag output to obtain a dataframe with time-series of F, Fmsy, SSB and MSY B trigger for
each stock in the Ecoregion, according to the last assessment (relative to the set year)
}
\note{
Can add some helpful information here
}
\examples{
\dontrun{
plot1 <- plot_CLD_bar(CLDstatus, guild="Demersal", return_data = TRUE)
}

}
\references{
The ICES stock information Database web sevices: \url{http://sid.ices.dk/services/}
}
\seealso{
\code{\link{plot_CLD_bar}} Stock status relative to reference points. 

\code{\link{icesFO-package}} gives an overview of the package.
}
