% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_discard_current.R
\name{plot_discard_current}
\alias{plot_discard_current}
\title{Returns a plot of discard rate by fish category for an ecoregion in the last assessment year.}
\usage{
plot_discard_current(x, year, caption = TRUE, cap_year, cap_month,
  return_data = FALSE)
}
\arguments{
\item{x}{a dataframe output of CLD_trends}

\item{year}{the year required}

\item{caption}{logical for whether there should be a caption}

\item{cap_year}{the year to be shown in the figure caption}

\item{cap_month}{the month to be shown in the figure caption, the accession date to SAG usually}

\item{return_data}{a parameter indicating if the data behind the plot should be returned as a dataframe}
}
\value{
A plot
}
\description{
Wrangling of format_sag output to obtain a dataframe with time-series of F, Fmsy, SSB and MSY B trigger for
each stock in the Ecoregion, according to the last assessment (relative to the set year)
}
\note{
Can add some helpful information here
}
\examples{
\dontrun{
plot1 <- plot_discard_trends(CLDtrends, caption = T,cap_year, cap_month, return_data = TRUE)
}

}
\references{
The ICES stock information Database web sevices: \url{http://sid.ices.dk/services/}
}
\seealso{
\code{\link{plot_discard_current}} a plot of landings and discards by fish category in the last year for an ecoregion. 

\code{\link{icesFO-package}} gives an overview of the package.
}
