% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ecoregion_map.R
\name{plot_ecoregion_map}
\alias{plot_ecoregion_map}
\title{Returns an map of the ecoregion and the divisions included in it}
\usage{
plot_ecoregion_map(ecoregion, ices_areas)
}
\arguments{
\item{ecoregion}{a simple features dataframe output of load_ecoregions( ... )}

\item{ices_areas}{a simple features dataframe output of load_areas( ... )}
}
\value{
A ggplot plot
}
\description{
Returns an map of the ecoregion and the divisions included in it
}
\examples{
\dontrun{
plot1 <- plot_ecoregionMap(ecoregions,areas, ecoregion = "Baltic Sea Ecoregion", 
return_data = FALSE)
}

}
\seealso{
\code{\link{plot_CLD_bar}} Stock status relative to reference points. 

\code{\link{icesFO-package}} gives an overview of the package.
}
