% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ecoregionMap.R
\name{plot_ecoregionMap}
\alias{plot_ecoregionMap}
\title{Returns an map of the ecoregion and the divisions included in it}
\usage{
plot_ecoregionMap(x, y, ecoregion)
}
\arguments{
\item{x}{a dataframe output of load_ecoregions.R}

\item{y}{a dataframe output of load_areas.R}

\item{ecoregion}{the ecoregion of interest}
}
\value{
A plot
}
\description{
Returns an map of the ecoregion and the divisions included in it
}
\note{
Can add some helpful information here
}
\examples{
\dontrun{
plot1 <- plot_ecoregionMap(ecoregions,areas, ecoregion = "Baltic Sea Ecoregion", 
return_data = FALSE)
}

}
\seealso{
\code{\link{plot_CLD_bar}} Stock status relative to reference points. 

\code{\link{icesFO-package}} gives an overview of the package.
}
