% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_GES_pies.R
\name{plot_GES_pies}
\alias{plot_GES_pies}
\title{Returns an ordered plot of catch bars colored according to
F/F<sub>MSY</sub> and SSB/MSY B<sub>trigger</sub>
by fish category and ecoregion}
\usage{
plot_GES_pies(
  x,
  y,
  cap_month = "August",
  cap_year = "2019",
  return_data = FALSE
)
}
\arguments{
\item{x}{a dataframe output of format_sag_status.R}

\item{y}{a dataframe output of stockstatus_CLD_current.R}

\item{cap_month}{the month to be shown in the figure caption, the accession date to SAG usually}

\item{cap_year}{the year to be shown in the figure caption}

\item{return_data}{a parameter indicating if the data behind the plot should be returned as a dataframe}
}
\value{
A plot
}
\description{
Wrangling of format_sag output to obtain a dataframe with time-series of F, Fmsy, SSB and MSY B trigger for
each stock in the Ecoregion, according to the last assessment (relative to the set year)
}
\note{
Can add some helpful information here
}
\examples{
\dontrun{
plot1 <- plot_CLD_bar(CLDstatus, guild="Demersal", return_data = TRUE)
}

}
\references{
The ICES stock information Database web sevices: \url{http://sid.ices.dk/services/}
}
\seealso{
\code{\link{plot_CLD_bar}} Stock status relative to reference points.

\code{\link{icesFO-package}} gives an overview of the package.
}
