% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_status_kobe.R
\name{plot_kobe}
\alias{plot_kobe}
\title{Returns a scatterplot of F/F<sub>MSY</sub> and SSB/MSY B<sub>trigger</sub>
by fish category and ecoregion}
\usage{
plot_kobe(x, guild, caption = F, cap_year, cap_month, return_data = T)
}
\arguments{
\item{x}{a dataframe output of stockstatus+catch_current.R}

\item{guild}{an identifier of the Fisheries guild to plot}

\item{cap_year}{the year to be shown in the figure caption}

\item{cap_month}{the month to be shown in the figure caption, the accession date to SAG usually}

\item{return_data}{a parameter indicating if the data behind the plot should be returned as a dataframe}
}
\value{
A plot
}
\description{
Wrangling of format_sag output to obtain a dataframe with time-series of F, Fmsy, SSB and MSY B trigger for
each stock in the Ecoregion, according to the last assessment (relative to the set year)
}
\note{
Can add some helpful information here
}
\examples{
\dontrun{
plot1 <- plot_kobe(CLDstatus, guild="Demersal", return_data = TRUE)
}

}
\references{
The ICES stock information Database web sevices: \url{http://sid.ices.dk/services/}
}
\seealso{
\code{\link{plot_kobe}} Stock status relative to reference points. 

\code{\link{icesFO-package}} gives an overview of the package.
}
