% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sar_map.R
\name{plot_sar_map}
\alias{plot_sar_map}
\title{Returns an map of the ecoregion and the divisions included in it}
\usage{
plot_sar_map(sar, ecoregion, what)
}
\arguments{
\item{sar}{a dataframe of swept area ratio data with a WKT 
columns (see notes)}

\item{ecoregion}{a dataframe output of load_ecoregion( ... )}
}
\value{
A ggplot object
}
\description{
Returns an map of the ecoregion and the divisions included in it
}
\note{
The effort data.frame is conveted to an sf object for plotting and
is expeted to have a Well Known Text column containing polygons
of c-squares
}
\examples{
\dontrun{
ecoregion <- load_ecoregion("Baltic Sea")
sar <- icesVMS::get_sar_map("Baltic Sea")
plot1 <- plot_effort_map(sar, ecoregion)
}

}
\seealso{
\code{\link{plot_ecoregion_map}} plots ICES areas and ecoregion.
}
