% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stecf.R
\name{plot_stecf}
\alias{plot_stecf}
\title{STECF Landings over time by country, guild, or species}
\usage{
plot_stecf(x, type, variable = NULL, cap_year, cap_month, line_count,
  stecf_report, return_data = FALSE)
}
\arguments{
\item{x}{a dataframe resulting from format_stecf_effort function}

\item{type}{need description}

\item{variable}{need description}

\item{cap_year}{the year to be shown in the figure caption}

\item{cap_month}{the month to be shown in the figure caption, the accession date to SAG usually}

\item{line_count}{indicates the number of lines to be shown in the graph}

\item{stecf_report}{need description}

\item{return_data}{a parameter indicating if the data behind the plot should be returned as a dataframe}
}
\value{
A ggplot2 object
}
\description{
The \code{stecf_plot} function returns an area or line plot of landings (historic and official catch) for an ecoregion by country
guild, or species.
}
\examples{
\dontrun{
stecf_plot("Greater North Sea Ecoregion", 
           metric = "EFFORT", 
           type = "GEAR", 
           return_plot = TRUE, 
           line_count = 4)
}
}
