% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_technical_interaction.R
\name{plot_technical_interactions}
\alias{plot_technical_interactions}
\title{Visualise a matrix of technical interactions.}
\usage{
plot_technical_interactions(x, col = heat.colors(5, rev = TRUE))
}
\arguments{
\item{x}{a square matrix with identical row and column names containing
values of tecnical interaction between species}

\item{col}{a vector of colours to use on the image plot}
}
\value{
A list with four elements, technical interactions for landings (recapLand)
        and catch (recapCatch) and tables of main gears involved in the interactions
        based on landings (MainGearsLandings) and catches (MainGearsCatches)
}
\description{
Based on a matrix of technical interactions produce an image plot
with a legend.
}
\examples{
\dontrun{
technical_interacton <- 
  compute_technical_interactions(STECF_landings, catchCoverage = 0.99)

plot_technical_interactions(technical_interacton$recapLand)
}

}
\seealso{
\code{\link{compute_technical_interactions}} for computing technical interactions.

\code{\link{icesFO-package}} gives an overview of the package.
}
