% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_VMS.R
\name{plot_vms}
\alias{plot_vms}
\title{VMS Effort and landings over time by country or gear}
\usage{
plot_vms(x, metric = NULL, type = NULL, cap_year, cap_month,
  line_count, return_data = FALSE)
}
\arguments{
\item{x}{a dataframe resulting from icesVMS::get_fo_effort() or icesVMS::get_fo_landings()}

\item{metric}{either "effort" or "landings"}

\item{type}{either "country" or "gear_category"}

\item{cap_year}{the year to be shown in the figure caption}

\item{cap_month}{the month to be shown in the figure caption, the accession date to SAG usually}

\item{line_count}{indicates the number of lines to be shown in the graph}

\item{return_data}{a parameter indicating if the data behind the plot should be returned as a dataframe}
}
\value{
A ggplot2 object
}
\description{
The \code{stecf_plot} function returns an area or line plot of landings (historic and official catch) for an ecoregion by country
guild, or species.
}
\examples{
\dontrun{
vms_plot("Greater North Sea Ecoregion", 
           metric = "effort", 
           type = "gear_category",
           line_count = 4)
}
}
