% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockstatus+catch_current.R
\name{stockstatus_CLD_current}
\alias{stockstatus_CLD_current}
\title{Stock status relative to reference points}
\usage{
stockstatus_CLD_current(x)
}
\arguments{
\item{x}{a dataframe output of format_sag function}
}
\value{
A data frame..
}
\description{
returns a data frame of stock status relative to reference points and
catch, discards, and landings by stock for the most recent assessment.
}
\note{
Can add some helpful information here
}
\examples{
\dontrun{
CLD_2018 <- catch_landings_discards(sag_formatted)
}

}
\references{
The ICES stock information Database web sevices: \url{http://sid.ices.dk/services/}
}
\seealso{
\code{\link{stock_trends}} for formatting data from the ICES Stock Assessment database. 

\code{\link{icesFO-package}} gives an overview of the package.
}
