% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_catches.R
\name{format_catches}
\alias{format_catches}
\title{Format the data from the time-series of ICES historical, official and preliminary catches.}
\usage{
format_catches(year, ecoregion, historical, official, preliminary = NULL,
  species_list, sid)
}
\arguments{
\item{year}{the year required}

\item{ecoregion}{an identifier of the Ecoregion of interest}

\item{historical}{a dataframe output from load_historical_catches() required.}

\item{official}{a dataframe output from load_official_catches() required.}

\item{preliminary}{a dataframe output from load_preliminary_catches() optional.}

\item{species_list}{a list of species common names and scientific names
from ASFIS}

\item{sid}{the main table from ICES Stock Information DB (SD)}
}
\value{
A data frame..
}
\description{
Format the data from the ICES Stock Assessment Database for the downloaded year and the specific Ecoregion
for which you are producing the Fisheries Overviews.
}
\note{
Can add some helpful information here
}
\examples{
\dontrun{
  sid <- load_sid(2019)
  species_list <- load_asfis_species()
  historical <- load_historical_catches()
  official <- load_official_catches()
  preliminary <- load_preliminary_catches(2018)
  catches_format <- 
    format_catches(2019, "Celtic Seas", historical, official, preliminary,
                   species_list, sid)
}

}
\references{
The ICES stock information Database web sevices: \url{http://sid.ices.dk/services/}
}
\seealso{
\code{\link{format_sid}} for formatting raw data from the ICES Stock Information database. 

\code{\link{icesFO-package}} gives an overview of the package.
}
