% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_catches.R
\name{format_catches}
\alias{format_catches}
\title{Format the data from the time-series of ICES historical, official and preliminary catches.}
\usage{
format_catches(year, ecoregion, x, y, z)
}
\arguments{
\item{ecoregion}{an identifier of the Ecoregion of interest}

\item{x}{a dataframe output from load_historical_catches() required.}

\item{y}{a dataframe output from load_official_catches() required.}

\item{z}{a dataframe output from load_preliminary_catches() optional.}
}
\value{
A data frame..
}
\description{
Format the data from the ICES Stock Assessment Database for the downloaded year and the specific Ecoregion
for which you are producing the Fisheries Overviews.
}
\note{
Can add some helpful information here
}
\examples{
\dontrun{
catches_format <- format_catches(x,y,z, "Celtic Seas")
}

}
\references{
The ICES stock information Database web sevices: \url{http://sid.ices.dk/services/}
}
\seealso{
\code{\link{format_sid}} for formatting raw data from the ICES Stock Information database. 

\code{\link{icesFO-package}} gives an overview of the package.
}
