% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icesFO-package.R
\docType{package}
\name{icesFO-package}
\alias{icesFO-package}
\alias{icesFO}
\title{icesFO: Functions to support the creation of ICES Fisheries Overviews}
\description{
Functions to support the creation of ICES Fisheries Overviews.

Functions to support the creation of ICES Fisheries Overviews, ....
}
\details{
\emph{Loading data:}
\tabular{ll}{
  \code{\link{load_sid}}                   \tab load data from SID\cr
  \code{\link{load_sag}}            \tab load data from SAG\cr
  \code{\link{load_catches}}      \tab load catch data from ICES\cr
}
\emph{Formatting datas:}
\tabular{ll}{
  \code{\link{format_sid}}      \tab format raw output from load_sid\cr
  \code{\link{format_stecf}}      \tab format raw output from load_sid\cr
}
\emph{Plots:}
\tabular{ll}{
  \code{\link{plot_CLD_bar}}      \tab plot\cr
  \code{\link{plot_GES_pies}}      \tab plot\cr
  \code{\link{plot_catch_trends}}      \tab plot\cr
  \code{\link{plot_discard_current}}      \tab plot\cr
  \code{\link{plot_discard_trends}}      \tab plot\cr
  \code{\link{plot_kobe}}      \tab plot\cr
  \code{\link{plot_status_prop_pies}}      \tab plot\cr
  \code{\link{plot_stecf}}      \tab plot\cr
  \code{\link{plot_stock_trends}}      \tab plot\cr
}
}
\author{
\strong{Maintainer}: Adriana Villamor \email{adriana.villamor@ices.dk}

Authors:
\itemize{
  \item Colin Millar
  \item Colm Lordan
  \item Youen Vermard
  \item Scott Large
}

Other contributors:
\itemize{
  \item Sarah Millar [contributor]
}


Adriana Villamor, Scott Large, Sarah Louise Millar and Colin Millar.
}
\keyword{internal#'}
