% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ecoregions.R
\name{load_ecoregions}
\alias{load_ecoregions}
\title{Download ecoregion polygons}
\usage{
load_ecoregions()
}
\value{
A simple features collection
}
\description{
Returns a simple features object with a polygon for each
ecoregion
}
\examples{
\dontrun{
  ecoregions <- load_ecoregions()
}

}
\seealso{
\code{\link{icesFO-package}} gives an overview of the package.
}
