% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sag.R
\name{load_sag}
\alias{load_sag}
\alias{load_sag_summary}
\alias{load_sag_refpts}
\alias{load_sag_status}
\title{Download data from the ICES Stock Assessment database.}
\usage{
load_sag_summary(year)

load_sag_refpts(year)

load_sag_status(year)
}
\arguments{
\item{year}{the year for which data is required.}
}
\value{
A data frame..
}
\description{
Download data from the ICES Stock Assessment database for the year in
which you are producing the Fisheries Overviews. Form SAG
}
\note{
Can add some helpful information here
}
\examples{
\dontrun{
sag_summary_raw <- load_sag_summary(2019)
sag_refpts_raw <- load_sag_refpts(2019)
sag_status_raw <- load_sag_status(2019)
}

}
\references{
The ICES stock assessment graphs Database web sevices: \url{http://standardgraphs.ices.dk/stockList.aspx}
}
\seealso{
\code{\link{load_sid}} for loading data from the ICES Stock Information database.

\code{\link{icesFO-package}} gives an overview of the package.
}
