% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_effort_map.R
\name{plot_effort_map}
\alias{plot_effort_map}
\title{Returns an map of the ecoregion and the divisions included in it}
\usage{
plot_effort_map(effort, ecoregion)
}
\arguments{
\item{effort}{a dataframe of effort data with a WKT columns (see notes)}

\item{ecoregion}{a dataframe output of load_ecoregion( ... )}
}
\value{
A ggplot object
}
\description{
Returns an map of the ecoregion and the divisions included in it
}
\note{
The effort data.frame is conveted to an sf object for plotting and
is expeted to have a Well Known Text column containing polygons
of c-squares
}
\examples{
\dontrun{
ecoregion <- load_ecoregion("Baltic Sea")
effort <- icesVMS::get_effort_map("Baltic Sea")

# convert to sf
effort$wkt <- sf::st_as_sfc(effort$wkt)
effort <- sf::st_sf(effort, sf_column_name = "wkt", crs = 4326)

# select gears to plot
gears <- c("Static", "Midwater", "Otter", "Demersal seine")
effort <- effort[effort$fishing_category_FO \%in\% gears,]

plot1 <- plot_effort_map(effort, ecoregion)
}

}
\seealso{
\code{\link{plot_ecoregion_map}} plots ICES areas and ecoregion.
}
