% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_effort_map.R
\name{plot_effort_map}
\alias{plot_effort_map}
\title{Returns an map of the ecoregion and the divisions included in it}
\usage{
plot_effort_map(effort, ecoregion_shape, ecoregion)
}
\arguments{
\item{effort}{a dataframe of effort data with a WKT columns (see notes)}

\item{ecoregion_shape}{a dataframe output of load_ecoregions()}

\item{ecoregion}{the ecoregion of interest (e.g. "Baltic Sea")}
}
\value{
A ggplot object
}
\description{
Returns an map of the ecoregion and the divisions included in it
}
\note{
The effort data.frame is conveted to an sf object for plotting and
is expeted to have a Well Known Text column containing polygons
of c-squares
}
\examples{
\dontrun{
ecoregions <- load_ecoregions()
effort <- icesVMS::get_effort_map("Baltic Sea")
plot1 <- 
  plot_effort_map(ecoregions, ecoregion = "Baltic Sea", 
                  return_data = FALSE)
}

}
\seealso{
\code{\link{plot_ecoregion_map}} plots ICES areas and ecoregion.
}
